# server/resource/vcenterws_db/Execution.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE Execution.js PLEASE EDIT ../custom/ExecutionCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_vcenterws_db = require('../../db/vcenterws_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/ExecutionCustom.js');

/*
 * SCHEMA DB Execution
 * 
	{
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		Script: {
			type: Schema.ObjectId,
			ref : "Execution"
		},
		user: {
			type: Schema.ObjectId,
			ref : "Execution"
		},
		
	}
 * 
 */



//CRUD METHODS


/**
 * ExecutionService.create
 *   @description CRUD ACTION create
 *
 */
app.post(properties.api + '/executions', function (req, res) {
	obj = new db_vcenterws_db.Execution(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * ExecutionService.delete
 *   @description CRUD ACTION delete
 *   @param ObjectId id Id
 *
 */
app['delete'](properties.api + '/executions/:id', function (req, res) {
	db_vcenterws_db.Execution.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * ExecutionService.findByScript
 *   @description CRUD ACTION findByScript
 *   @param Objectid key Id della risorsa Script da cercare
 *
 */
app.get(properties.api + '/executions/findByScript/:key', function (req, res) {

	db_vcenterws_db.Execution.find({ 'Script' : req.params.key}).exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
	
});

/**
 * ExecutionService.findByuser
 *   @description CRUD ACTION findByuser
 *   @param Objectid key Id della risorsa user da cercare
 *
 */
app.get(properties.api + '/executions/findByuser/:key', function (req, res) {

	db_vcenterws_db.Execution.find({ 'user' : req.params.key}).exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
	
});

/**
 * ExecutionService.get
 *   @description CRUD ACTION get
 *   @param ObjectId id Id 
 *
 */
app.get(properties.api + '/executions/:id', function (req, res) {
	db_vcenterws_db.Execution.findOne({_id:req.params.id}).exec(function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * ExecutionService.list
 *   @description CRUD ACTION list
 *
 */
app.get(properties.api + '/executions', function (req, res) {
	db_vcenterws_db.Execution.find().exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
});

/**
 * ExecutionService.update
 *   @description CRUD ACTION update
 *   @param ObjectId id Id
 *
 */
app.post(properties.api + '/executions/:id', function (req, res) {
	db_vcenterws_db.Execution.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */